feat: generate macros for components instead of structs
Generate macros to create components instead of generating structs.
And use that macro to call components inside another components.
