+++ origami-engine/src/lib.rs
@@ -1,15 +1,15 @@
-pub use origami_macros::og;
-
-pub trait Layout<T> {
-    fn extend(&self, s: &mut String, blocks: &T);
-    fn extend_with_childrens(&self, s: &mut String, blocks: &T, childrens: impl Fn(&mut String));
-}
-
-pub trait Origami {
-    fn to_html(&self) -> String;
-    fn push_html_to_string(&self, s: &mut String);
-    fn push_html_to_string_with_childrens(&self, s: &mut String, childrens: impl Fn(&mut String));
-}
+pub use origami_macros::{anon, comp};
 
 #[cfg(feature = "html_escape")]
 pub use html_escape::encode_text_to_string;
+
+pub struct Origami(pub String);
+
+#[cfg(feature = "axum")]
+use ::axum::response::{Html, IntoResponse, Response};
+#[cfg(feature = "axum")]
+impl IntoResponse for Origami {
+    fn into_response(self) -> Response {
+        Html(self.0).into_response()
+    }
+}
