+++ origami-macros/Cargo.toml
@@ -16,6 +16,5 @@ quote = "1.0.36"
 syn = { version = "2.0.72", features = ["extra-traits", "full"] }
 
 [features]
-default = ["html_escape", "axum"]
+default = ["html_escape"]
 html_escape = []
-axum = []
