+++ origami-macros/src/lib.rs
@@ -2,16 +2,27 @@ use proc_macro::TokenStream;
 use quote::quote;
 use syn::parse_macro_input;
 
-use self::origami::Origami;
+use self::anon::Anon;
+use self::comp::Component;
 
-mod origami;
+mod anon;
+mod comp;
 mod utils;
 
 #[proc_macro]
-pub fn og(input: TokenStream) -> TokenStream {
-    let origami = parse_macro_input!(input as Origami);
+pub fn comp(input: TokenStream) -> TokenStream {
+    let comp = parse_macro_input!(input as Component);
     quote! {
-        #origami
+        #comp
+    }
+    .into()
+}
+
+#[proc_macro]
+pub fn anon(input: TokenStream) -> TokenStream {
+    let anon = parse_macro_input!(input as Anon);
+    quote! {
+        #anon
     }
     .into()
 }
