+++ origami-macros/src/utils.rs
@@ -1,10 +1,7 @@
 pub mod kw {
     use syn::custom_keyword;
-    custom_keyword!(block);
-    custom_keyword!(extend);
-    custom_keyword!(layout);
+    custom_keyword!(no_s);
     custom_keyword!(size_hint);
-    custom_keyword!(childrens);
     custom_keyword!(include);
     custom_keyword!(escape);
     custom_keyword!(noescape);
